Migrating to vue2 in favour of libs support
