+++ flint.ui/src/App.vue
@@ -1,5 +1,7 @@
 <template>
-  <div id="app">
-    <router-view />
-  </div>
+  <router-view></router-view>
 </template>
+
+<script>
+export default {};
+</script>
