+++ flint.ui/src/assets/styles/tailwind.css
@@ -1303,11 +1303,6 @@ select {
   background-color: rgba(0, 0, 0, var(--tw-bg-opacity));
 }
 
-.hover\:bg-blue-500:hover {
-  --tw-bg-opacity: 1;
-  background-color: rgba(59, 130, 246, var(--tw-bg-opacity));
-}
-
 .hover\:bg-green-200:hover {
   --tw-bg-opacity: 1;
   background-color: rgba(187, 247, 208, var(--tw-bg-opacity));
@@ -1342,9 +1337,14 @@ select {
   background-image: linear-gradient(to right, var(--tw-gradient-stops));
 }
 
-.from-pink-500 {
-  --tw-gradient-from: #ec4899;
-  --tw-gradient-stops: var(--tw-gradient-from), var(--tw-gradient-to, rgba(236, 72, 153, 0));
+.bg-gradient-to-tl {
+  background-image: -webkit-gradient(linear, right bottom, left top, from(var(--tw-gradient-stops)));
+  background-image: linear-gradient(to top left, var(--tw-gradient-stops));
+}
+
+.from-green-300 {
+  --tw-gradient-from: #86efac;
+  --tw-gradient-stops: var(--tw-gradient-from), var(--tw-gradient-to, rgba(134, 239, 172, 0));
 }
 
 .from-green-400 {
@@ -1352,12 +1352,16 @@ select {
   --tw-gradient-stops: var(--tw-gradient-from), var(--tw-gradient-to, rgba(74, 222, 128, 0));
 }
 
-.to-blue-500 {
-  --tw-gradient-to: #3b82f6;
+.via-blue-500 {
+  --tw-gradient-stops: var(--tw-gradient-from), #3b82f6, var(--tw-gradient-to, rgba(59, 130, 246, 0));
 }
 
-.to-yellow-500 {
-  --tw-gradient-to: #eab308;
+.to-purple-600 {
+  --tw-gradient-to: #9333ea;
+}
+
+.to-blue-500 {
+  --tw-gradient-to: #3b82f6;
 }
 
 .bg-center {
@@ -2000,6 +2004,10 @@ select {
   height: 1.5rem;
 }
 
+.h-10 {
+  height: 2.5rem;
+}
+
 .h-12 {
   height: 3rem;
 }
@@ -2112,11 +2120,6 @@ select {
   margin: -6rem;
 }
 
-.mx-2 {
-  margin-left: 0.5rem;
-  margin-right: 0.5rem;
-}
-
 .my-4 {
   margin-top: 1rem;
   margin-bottom: 1rem;
@@ -2137,6 +2140,10 @@ select {
 
 .mt-1 {
   margin-top: 0.25rem;
+}
+
+.mb-1 {
+  margin-bottom: 0.25rem;
 }
 
 .mt-2 {
@@ -2171,10 +2178,6 @@ select {
   margin-bottom: 1.25rem;
 }
 
-.mr-6 {
-  margin-right: 1.5rem;
-}
-
 .mb-6 {
   margin-bottom: 1.5rem;
 }
@@ -2183,16 +2186,8 @@ select {
   margin-top: 2rem;
 }
 
-.mr-8 {
-  margin-right: 2rem;
-}
-
 .mt-10 {
   margin-top: 2.5rem;
-}
-
-.mb-12 {
-  margin-bottom: 3rem;
 }
 
 .mr-auto {
@@ -2270,11 +2265,6 @@ select {
 
 .opacity-75 {
   opacity: 0.75;
-}
-
-.outline-none {
-  outline: 2px solid transparent;
-  outline-offset: 2px;
 }
 
 .focus\:outline-none:focus {
@@ -2381,10 +2371,6 @@ select {
   padding: 1rem;
 }
 
-.p-5 {
-  padding: 1.25rem;
-}
-
 .p-6 {
   padding: 1.5rem;
 }
@@ -2448,10 +2434,6 @@ select {
   padding-top: 0.25rem;
 }
 
-.pt-2 {
-  padding-top: 0.5rem;
-}
-
 .pt-4 {
   padding-top: 1rem;
 }
@@ -2486,6 +2468,31 @@ select {
 
 .pb-32 {
   padding-bottom: 8rem;
+}
+
+.placeholder-gray-600::-webkit-input-placeholder {
+  --tw-placeholder-opacity: 1;
+  color: rgba(82, 82, 91, var(--tw-placeholder-opacity));
+}
+
+.placeholder-gray-600::-moz-placeholder {
+  --tw-placeholder-opacity: 1;
+  color: rgba(82, 82, 91, var(--tw-placeholder-opacity));
+}
+
+.placeholder-gray-600:-ms-input-placeholder {
+  --tw-placeholder-opacity: 1;
+  color: rgba(82, 82, 91, var(--tw-placeholder-opacity));
+}
+
+.placeholder-gray-600::-ms-input-placeholder {
+  --tw-placeholder-opacity: 1;
+  color: rgba(82, 82, 91, var(--tw-placeholder-opacity));
+}
+
+.placeholder-gray-600::placeholder {
+  --tw-placeholder-opacity: 1;
+  color: rgba(82, 82, 91, var(--tw-placeholder-opacity));
 }
 
 .pointer-events-none {
@@ -3353,6 +3360,12 @@ select {
         flex-wrap: nowrap;
   }
 
+  .md\:items-center {
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+  }
+
   .md\:items-stretch {
     -webkit-box-align: stretch;
         -ms-flex-align: stretch;
@@ -3369,6 +3382,16 @@ select {
     -webkit-box-pack: justify;
         -ms-flex-pack: justify;
             justify-content: space-between;
+  }
+
+  .md\:flex-grow {
+    -webkit-box-flex: 1;
+        -ms-flex-positive: 1;
+            flex-grow: 1;
+  }
+
+  .md\:mb-0 {
+    margin-bottom: 0px;
   }
 
   .md\:mt-4 {
@@ -3454,6 +3477,10 @@ select {
     width: 16rem;
   }
 
+  .md\:w-2\/3 {
+    width: 66.666667%;
+  }
+
   .md\:w-4\/12 {
     width: 33.333333%;
   }
@@ -3496,10 +3523,6 @@ select {
 
   .xl\:w-6\/12 {
     width: 50%;
-  }
-
-  .xl\:w-8\/12 {
-    width: 66.666667%;
   }
 }
