+++ flint.ui/src/components/Cards/CardInfoConfig.vue
@@ -37,18 +37,12 @@
       </p>
       </div>
     </div>
-    <Alert
-    v-show="alertOpen"
-    v-bind:alertText="{cardAlertConfirmation}.cardAlertConfirmation"
-    ></Alert>
   </div>
 </template>
 <script>
 
-// const axios = require('axios')
 import ConfirmRun from "@/components/Prompts/ConfirmRun";
 import ConfirmConfig from "@/components/Prompts/ConfirmConfig";
-import Alert from "@/components/Alerts/Alert";
 
 export default {
   name: "card-info-config",
@@ -77,21 +71,15 @@ export default {
       type: String, 
       default: "spec",
     },
-    cardAlertConfirmation: {
-      type: String, 
-      default: "___ route ran successfully",
-    },
   },
   components: {
     ConfirmRun,
-    ConfirmConfig,
-    Alert,
+    ConfirmConfig
   },
   data() {
     return {
       isConfirmRunModalVisible: false,
-      alertOpen: false,
-      isConfirmConfigModalVisible: false,
+      isConfirmConfigModalVisible: false
     };
   },
   methods: {
@@ -100,12 +88,6 @@ export default {
     },
     closeConfirmConfigModal() {
       this.isConfirmConfigModalVisible = false;
-    },
-    delayAlert() {
-      this.alertOpen = true
-      setTimeout(()=> {
-        this.alertOpen = false
-      }, 2500)
     },
     showConfirmRunModal() {
       this.isConfirmRunModalVisible = true;
@@ -130,23 +112,17 @@ export default {
   
       if (api_route == "point") {
         this.apiRoute_point()
-        this.delayAlert()
       }
       else if (api_route == "rothc") {
         this.apiRoute_rothc()
-        this.delayAlert()
       }
       else 
         this.apiRoute_nonexistent()
     },
 
     apiRoute_point() {
-      //sending the new config
-      console.log('POINT route invoked with new configs')      
-      // axios.post('http://127.0.0.1:8080/point')
-      //   .then(response => console.log(response))
-      //   .catch(error => console.log(error))
-      // return this.title
+      //sending the new point config
+      console.log('POINT route invoked with new configs')
     },
     apiRoute_rothc() {
       // sending the new rothc config
