+++ flint.ui/src/components/Headers/ExampleCardsConfig.vue
@@ -11,7 +11,6 @@
               cardFunctionRun="Run"
               cardFunctionConfig="Configure"
               cardMethodName="point"
-              cardAlertConfirmation="Point route ran successfully"
             />
           </div>
           <div class="w-full lg:w-6/12 xl:w-3/12 px-4 content-center">
@@ -22,7 +21,6 @@
               cardFunctionRun="Run"
               cardFunctionConfig="Configure"
               cardMethodName="rothc"
-              cardAlertConfirmation="RothC route ran successfully"
             />
           </div>
         </div>
