+++ flint.ui/src/components/Sliders/ConfigSliders.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-  import Slider from '@vueform/slider'
+import Slider from "@vueform/slider/dist/slider.vue2.js";
   
   export default {
     components: {
