+++ flint.ui/src/layouts/Flint.vue
@@ -4,7 +4,6 @@
     <div class="relative md:ml-64 bg-blueGray-100">
       <div class="px-4 md:px-10 mx-auto w-full -m-24 h-screen">
         <router-view />
-        <footer-admin />
       </div>
     </div>
   </div>
