+++ flint.ui/src/store/index.js
@@ -1,7 +1,10 @@
-import { createStore } from "vuex";
+import Vue from "vue";
+import Vuex from "vuex";
 import axios from "axios";
 
-export default createStore({
+Vue.use(Vuex);
+
+export default new Vuex.Store({
   state: {
     Point_config: {
       LocalDomain: {
@@ -547,8 +550,14 @@ actions: {
     );
     axios
       .post("http://127.0.0.1:8080/point", final_FLINT_config_string)
-      .then((response) => console.log(response))
-      .catch((error) => console.log(error));
+      .then((response) => {
+        Vue.$toast.success(`${response}`, { timeout: 2000 });
+        console.log(response);
+        })
+      .catch(error => {
+        Vue.$toast.error(`${error}`, { timeout: 2000 });
+        console.log(error);
+      })
   },
 
   send_rothcConfig() {
@@ -564,12 +573,14 @@ actions: {
     );
     axios
       .post("http://127.0.0.1:8080/rothc", final_RothC_config_string)
-      .then((response) => console.log(response))
-      .catch((error) => console.log(error));
-    },
-},
-modules: {
-
+      .then((response) => {
+        Vue.$toast.success(`${response}`, { timeout: 2000 });
+        console.log(response);
+      })
+      .catch(error => {
+        Vue.$toast.error(`${error}`, { timeout: 2000 });
+        console.log(error);
+      });
+  },
 }
-
 });
