+++ flint.ui/src/views/Landing.vue
@@ -60,7 +60,6 @@
           </div>
         </div>
       </div>
-    <footer-component />
   </div>
 </template>
 <script>
