+++ flint.ui/src/views/flint/ConfigurationsPoint.vue
@@ -62,11 +62,13 @@
 </template>
 
 <script>
-import Slider from "@vueform/slider";
+import Slider from "@vueform/slider/dist/slider.vue2.js";
+// import Slider from "@vueform/slider";
+// import Slider from "@vueform/slider/dist/slider.vue2.js";
 
 export default {
   components: {
-    Slider,
+    Slider
   },
   data() {
     return {
@@ -118,7 +120,6 @@ export default {
 };
 </script>
 
-<style src="@vueform/slider/themes/default.css"></style>
 
 <style>
 .slider-component {
