+++ flint.ui/src/views/flint/ConfigurationsRothc.vue
@@ -51,9 +51,7 @@
 import Rothctemplate from "@/views/flint/Rothctemplate.vue"
 
 export default {
-
-  data() {
-  },
+  
   components: {
     Rothctemplate,
   },
