+++ flint.ui/src/views/flint/configurations/rothc/RothcAvgairtemp.vue
@@ -95,10 +95,5 @@ export default {
       }
     },
   },
-  components: {
-  },
-  data() {
-  },     
-  methods: {
-  }}
+}
 </script>
