+++ flint.ui/src/views/flint/configurations/rothc/RothcInitsoil.vue
@@ -204,12 +204,5 @@ export default {
       }
     },
   },
-  components: {
-  },
-  data() {
-      return {
-      }},
-       
-  methods: {
-  }}
+}
 </script>
