+++ flint.ui/src/views/flint/configurations/rothc/RothcOpenpanevap.vue
@@ -95,10 +95,5 @@ export default {
       }
     },
   },
-  components: {
-  },
-  data() {
-  },     
-  methods: {
-  }}
+}
 </script>
