+++ flint.ui/src/views/flint/configurations/rothc/RothcPrescm.vue
@@ -92,11 +92,5 @@ export default {
       }
     },
   },
-  components: {
-    
-  },
-  data() {
-  },    
-  methods: {
-  }}
+}
 </script>
