+++ flint.ui/src/views/flint/configurations/rothc/RothcRainfall.vue
@@ -95,12 +95,5 @@ export default {
       }
     },
   },
-  components: {
-    
-  },
-  data() {
-  },
-  methods: {
-    
-  }}
+}
 </script>
