+++ flint.ui/src/views/flint/configurations/rothc/RothcSoil.vue
@@ -427,9 +427,5 @@ export default {
       }
     },
   },
-  components: {
-    
-  },
-  methods: {
-  }}
+}
 </script>
